refactor: add types to appView.ts
